refactore visualisaton plugin
