+++ package.json
@@ -1,10 +1,10 @@
 {
-  "name": "kbn-vis-multigraph",
-  "version": "0.0.0",
+  "name": "kbn_vis_multiple_graph",
+  "version": "0.0.1",
   "description": "kibana test plugin",
-  "main": "index.js",
+  "main": "kbn_vis_multiple_graph.js",
   "kibana": {
-    "version": "6.X",
+    "version": "6.1.3",
     "templateVersion": "7.2.4"
   },
   "scripts": {
@@ -31,7 +31,7 @@
     "lodash": "^4.4.0",
     "expect.js": "^0.3.1"
   },
-  "dependencies" : {
-    "c3": "0.4.17"
+  "dependencies": {
+    "c3": "^0.4.18"
   }
 }
