+++ public/kbn_vis_multiple_graph.html
