+++ public/kbn_vis_multiple_graph_params.html
