+++ public/less/kbn_vis_multiple_graph.less
@@ -1,0 +1,18 @@
+
+.c3-vis {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-around;
+  align-items: center;
+  align-content: space-around;
+
+  .c3-container {
+    text-align: center;
+    padding: 1em;
+    width:100%;
+    position:relative;
+  }
+
+}
